# wiki/summaries/OAS-EDU-V1.yaml
openapi: "3.0.3"
info:
  title: "EDU 培训机构教务收费管理系统 API"
  version: "1.0.0"
  description: |
    基于 ASD-EDU-V1（模块化分层单体）、ADR-002（接口契约决策）的 OpenAPI 3.0.3 接口规范。
    覆盖 7 模块核心 API，统一响应体 `{code, message, data, timestamp, traceId}`。
  contact:
    name: "EDU 项目组"
servers:
  - url: "https://api.edu.example.com/api/v1"
    description: "生产环境（HTTPS）"
  - url: "http://localhost:8080/api/v1"
    description: "本地开发环境"
...
paths:
...
  /crm/leads:
    get:
      tags: [CRM]
      summary: "查询线索列表"
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, FOLLOWING, CONVERTED, LOST] }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200": { $ref: "#/components/responses/PaginatedLeads" }
    post:
      tags: [CRM]
      summary: "创建线索（多渠道自动建档或手工登记）"
      requestBody: { $ref: "#/components/requestBodies/CreateLead" }
      responses:
        "201": { $ref: "#/components/responses/LeadCreated" }
        "409": { $ref: "#/components/responses/DuplicateLead" }
  /crm/leads/{leadId}/convert:
    post:
      tags: [CRM]
      summary: "线索转化为正式学员"
      parameters:
        - name: leadId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                courseIds: { type: array, items: { type: integer } }
                assignedConsultantId: { type: integer }
      responses:
        "200": { $ref: "#/components/responses/StudentCreated" }
...
  /enrollment/orders:
    post:
      tags: [Enrollment]
      summary: "创建报名订单"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, classId]
              properties:
                studentId: { type: integer }
                classId: { type: integer }
                couponCode: { type: string }
                discountReason: { type: string }
      responses:
        "201": { $ref: "#/components/responses/OrderCreated" }
        "400": { $ref: "#/components/responses/BusinessError" }
  /enrollment/orders/{orderId}:
    get:
      tags: [Enrollment]
      summary: "查询订单详情（含支付状态）"
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/OrderDetail" }
  /enrollment/orders/{orderId}/payment:
    post:
      tags: [Enrollment]
      summary: "发起支付（生成微信/支付宝支付参数）"
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel: { type: string, enum: [WECHAT, ALIPAY] }
      responses:
        "200": { $ref: "#/components/responses/PaymentInitiated" }
  /enrollment/orders/{orderId}/cancel:
    post:
      tags: [Enrollment]
      summary: "取消订单（仅待支付状态可取消）"
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/OrderCancelled" }

  # ── Academic: 教务排课与消课管理 ──
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Access Token（2h 有效期）"
  responses:
    PaginatedLeads:
      description: "分页线索列表"
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      content: { type: array, items: { $ref: "#/components/schemas/Lead" } }
                      page: { type: integer }
                      size: { type: integer }
                      totalElements: { type: integer }
                      totalPages: { type: integer }
    LeadCreated:
      description: "线索创建成功"
    DuplicateLead:
      description: "线索重复（手机号已存在）"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    StudentCreated:
      description: "学员创建成功"
...
    OrderCreated:
      description: "订单创建成功"
    OrderDetail:
      description: "订单详情（含支付状态）"
    PaymentInitiated:
      description: "支付参数（二维码/链接）"
    OrderCancelled:
      description: "订单已取消"
...
    BusinessError:
      description: "业务错误"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  requestBodies:
    CreateLead:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [name, phone]
            properties:
              name: { type: string, description: "学员/家长姓名" }
              phone: { type: string, pattern: "^1[3-9]\\d{9}$", description: "手机号（将自动查重）" }
              source: { type: string, enum: [ONLINE_AD, REFERRAL, WALK_IN, GROUND_PUSH], description: "线索来源" }
              campusId: { type: integer, description: "归属校区" }
  schemas:
    ApiResponse:
      type: object
      required: [code, message, timestamp, traceId]
      properties:
        code: { type: integer, description: "0=成功，1xxxxx=业务异常，9xxxxx=系统异常" }
        message: { type: string }
        data: { type: object }
        timestamp: { type: string, format: date-time }
        traceId: { type: string }
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data: { type: object, nullable: true, default: null }
          example:
            code: 10101
            message: "手机号格式错误"
            data: null
            timestamp: "2026-06-24T10:00:00Z"
            traceId: "a1b2c3d4e5f6"
    Lead:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string, description: "中间4位脱敏" }
        source: { type: string }
        status: { type: string, enum: [PENDING, FOLLOWING, CONVERTED, LOST] }
        createdAt: { type: string, format: date-time }
...
security:
  - BearerAuth: []